feat: Add a seperate docker image for production
We the .env file in dev and can be used in prod too but in severless environments
the .env file needs to provisioned meanwhile it's possible to just pass env vars
to most serverless runners.

Also for obvious reasons the prod setup does not need everything in the dev env.

Signed-off-by: nfebe <fenn25.fn@gmail.com>